chore: added tests and refactor validation
